Only compile Crossdock support when it's actually requested
Build Crossdock support into the UnitTest executable not the library, and
only build it when JAEGERTRACING_BUILD_CROSSDOCK is set.

In addition to shrinking the library this lets us drop the exact Thrift 0.9.2
or 0.9.3 requirement when not building with Crossdock. The IDL bug / Thrift bug
discussed in #45 only affects the IDL used with Crossdock testing.

Or at least, it would let us drop the Thrift requirement if not for the problems
with std::shared_ptr vs boost::shared_ptr.

Signed-off-by: Craig Ringer <craig@2ndquadrant.com>